requirements.txt
astroid==2.3.3
autopep8==1.5
boto3==1.12.26
botocore==1.15.26
click==7.1.1
colorama==0.4.3
docutils==0.15.2
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.12.0
s3transfer==0.3.3
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
